refactor(import): use new import form in game edit page
