+++ frontend/src/games/import/ImportWizard.tsx
@@ -1,0 +1,82 @@
+import { useState } from 'react';
+
+import { TabContext, TabPanel } from '@mui/lab';
+import { Box, Tab, Tabs } from '@mui/material';
+import { RemoteGame } from '../../api/gameApi';
+import { OnlineGameForm } from './OnlineGameForm';
+import { PGNForm } from './PGNForm';
+import { PositionForm } from './PositionForm';
+
+enum ImportSource {
+    URL = 'url',
+    Position = 'position',
+    PGNText = 'pgn-text',
+}
+
+interface ImportTabPanelProps {
+    children: React.ReactNode;
+    source: ImportSource;
+}
+
+const ImportTabPanel: React.FC<ImportTabPanelProps> = ({ children, source }) => {
+    return (
+        <TabPanel value={source} sx={{ px: { xs: 0, sm: 3 } }}>
+            {children}
+        </TabPanel>
+    );
+};
+
+interface ImportWizardProps {
+    loading: boolean;
+    onSubmit: (game: RemoteGame) => void;
+}
+
+export const ImportWizard: React.FC<ImportWizardProps> = ({ onSubmit, loading }) => {
+    const [source, setSource] = useState<ImportSource>(ImportSource.URL);
+
+    const handleTabChange = (_event: React.SyntheticEvent, source: ImportSource) => {
+        setSource(source);
+    };
+
+    return (
+        <>
+            <TabContext value={source}>
+                <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
+                    <Tabs
+                        onChange={handleTabChange}
+                        value={source}
+                        aria-label='import sources tabs'
+                        variant='scrollable'
+                    >
+                        <Tab
+                            value={ImportSource.URL}
+                            label={'Lichess & Chess.com'}
+                            data-cy={'import-url'}
+                        />
+                        <Tab
+                            value={ImportSource.Position}
+                            label='Position'
+                            data-cy={'import-position'}
+                        />
+                        <Tab
+                            value={ImportSource.PGNText}
+                            label='PGN'
+                            data-cy={'import-pgn-text'}
+                        />
+                    </Tabs>
+                </Box>
+                <ImportTabPanel source={ImportSource.URL}>
+                    <OnlineGameForm onSubmit={onSubmit} loading={loading} />
+                </ImportTabPanel>
+                <ImportTabPanel source={ImportSource.PGNText}>
+                    <PGNForm onSubmit={onSubmit} loading={loading} />
+                </ImportTabPanel>
+                <ImportTabPanel source={ImportSource.Position}>
+                    <PositionForm onSubmit={onSubmit} loading={loading} />
+                </ImportTabPanel>
+            </TabContext>
+        </>
+    );
+};
+
+export default ImportWizard;
